requirements.txt
pyobjc-core; platform_system == "Darwin"
pyobjc; platform_system == "Darwin"
pyobjc-framework-Quartz; platform_system == "Darwin"
pyobjc-framework-Vision; platform_system == "Darwin"
pyobjc-framework-Cocoa; platform_system == "Darwin"
Pillow
pyautogui
psutil
rich
loguru
pytesseract
fastapi
uvicorn
transformers
torch